Move Jupyter directive to its own file
This is so we can have unconditional imports at the top level of the
file without worrying about installed packages.
